feat: move log structs to models package
